# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute Commerce Platform API
  description: |
    Enterprise-grade B2B FMCG commerce platform combining multi-channel ordering,
    embedded finance, and gig economy logistics.
  version: 1.0.0
  contact:
    name: OmniRoute Team
    email: api@omniroute.io
  license:
    name: Proprietary
servers:
  - url: https://api.omniroute.io/v1
    description: Production
  - url: https://staging-api.omniroute.io/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local Development
...
paths:
...
  /workers:
    get:
      tags: [Workers]
      summary: List workers
      operationId: listWorkers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, busy, offline]
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: List of workers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerList'
    post:
      tags: [Workers]
      summary: Register a new worker
      operationId: createWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkerRequest'
      responses:
        '201':
          description: Worker created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      operationId: listTasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  # Notifications
...
components:
  parameters:
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    PageOffset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
  schemas:
...
    Worker:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [available, busy, offline]
        type:
          type: string
          enum: [delivery, sales, collection]
        rating:
          type: number
    WorkerList:
      type: object
      properties:
        workers:
          type: array
          items:
            $ref: '#/components/schemas/Worker'
        total:
          type: integer
    CreateWorkerRequest:
      type: object
      required: [first_name, last_name, phone, type]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        type:
          type: string
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
        priority:
          type: integer
        assigned_worker_id:
          type: string
          format: uuid
    TaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    CreateTaskRequest:
      type: object
      required: [type, order_id]
      properties:
        type:
          type: string
        order_id:
          type: string
          format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []